+++ src/extension/Services/CodeActionProvider_General.ts
@@ -6,6 +6,7 @@ import { CodeActionProviderExtractLabel } from './CodeActionProviderExtractLabel
 import { CodeActionProviderRefactorToValidate } from './CodeActionProviderRefactorToValidate';
 import { CodeActionProviderLocalVariableToGlobal } from './CodeActionProviderLocalVariableToGlobal';
 import { CodeActionProvider, TextDocument, Range, CodeActionContext, CancellationToken, CodeAction } from 'vscode';
+import { CodeActionProviderModifyProcedure } from './CodeActionProviderModifyProcedure';
 
 export class CodeActionProvider_General implements CodeActionProvider {
     async provideCodeActions(document: TextDocument, range: Range, context: CodeActionContext, token: CancellationToken): Promise<CodeAction[] | undefined> {
@@ -15,6 +16,7 @@ export class CodeActionProvider_General implements CodeActionProvider {
         myCodeActionProviders.push(new CodeActionProviderExtractLabel(document, range));
         myCodeActionProviders.push(new CodeActionProviderRefactorToValidate(document, range));
         myCodeActionProviders.push(new CodeActionProviderLocalVariableToGlobal(document, range));
+        myCodeActionProviders.push(new CodeActionProviderModifyProcedure(document, range));
 
         let codeActions: CodeAction[] = [];
 
@@ -30,7 +32,7 @@ export class CodeActionProvider_General implements CodeActionProvider {
                 let newActions: CodeAction[] = await myCodeActionProviderToExecute.createCodeActions();
                 codeActions = codeActions.concat(newActions);
             }
-            for(const codeAction of codeActions)
+            for (const codeAction of codeActions)
                 codeAction.title += ' (AL CodeActions)';
         }
         return codeActions;